feat: token dropdown logic poolPage
