+++ lib/pages/DesktopTrade.dart
@@ -1,5 +1,6 @@
 import 'package:ax_dapp/service/Athlete.dart';
 import 'package:ax_dapp/service/AthleteList.dart';
+import 'package:ax_dapp/service/AthleteTokenList.dart';
 import 'package:ax_dapp/service/Controller/Swap/AXT.dart';
 import 'package:ax_dapp/service/Controller/Swap/MATIC.dart';
 import 'package:ax_dapp/service/Controller/Swap/SXT.dart';
@@ -62,7 +63,7 @@ class _DesktopTradeState extends State<DesktopTrade> {
             onPressed: () {},
             child: Text(
               "Select token to swap with",
-              style: textStyle(Colors.blue, 16, true, false),
+              style: textStyle(Colors.blue, 16, true),
             )));
 
     if (tkn1 != null && tkn2 != null)
@@ -86,7 +87,7 @@ class _DesktopTradeState extends State<DesktopTrade> {
               },
               child: Text(
                 "Swap",
-                style: textStyle(Colors.black, 16, true, false),
+                style: textStyle(Colors.black, 16, true),
               )));
 
     return Container(
@@ -104,14 +105,14 @@ class _DesktopTradeState extends State<DesktopTrade> {
                     width: wid - 50,
                     alignment: Alignment.centerLeft,
                     child: Text("Swap",
-                        style: textStyle(Colors.white, 16, false, false))),
+                        style: textStyle(Colors.white, 16, false))),
                 Column(children: <Widget>[
                   //From text
                   Container(
                     width: wid - 75,
                     alignment: Alignment.centerLeft,
                     child: Text("From",
-                        style: textStyle(Colors.grey[400]!, 12, false, false)),
+                        style: textStyle(Colors.grey[400]!, 12, false)),
                   ),
                   // From-dropdown
                   Container(
@@ -151,8 +152,8 @@ class _DesktopTradeState extends State<DesktopTrade> {
                                                     style: textStyle(
                                                         Colors.grey[400]!,
                                                         8,
-                                                        false,
-                                                        false)))),
+                                                        false
+                                                        )))),
                                         SizedBox(
                                           width: 70,
                                           child: TextFormField(
@@ -160,13 +161,12 @@ class _DesktopTradeState extends State<DesktopTrade> {
                                               fromAmount = double.parse(value);
                                             },
                                             style: textStyle(Colors.grey[400]!,
-                                                22, false, false),
+                                                22, false),
                                             decoration: InputDecoration(
                                               hintText: '0.00',
                                               hintStyle: textStyle(
                                                   Colors.grey[400]!,
                                                   22,
-                                                  false,
                                                   false),
                                               contentPadding:
                                                   const EdgeInsets.all(9),
@@ -208,7 +208,7 @@ class _DesktopTradeState extends State<DesktopTrade> {
                     alignment: Alignment.centerLeft,
                     child: Text(
                       "To",
-                      style: textStyle(Colors.grey[400]!, 12, false, false),
+                      style: textStyle(Colors.grey[400]!, 12, false),
                     ),
                   ),
                   // To-dropdown
@@ -233,11 +233,11 @@ class _DesktopTradeState extends State<DesktopTrade> {
                                     toAmount = double.parse(value);
                                   },
                                   style: textStyle(
-                                      Colors.grey[400]!, 22, false, false),
+                                      Colors.grey[400]!, 22, false),
                                   decoration: InputDecoration(
                                     hintText: '0.00',
                                     hintStyle: textStyle(
-                                        Colors.grey[400]!, 22, false, false),
+                                        Colors.grey[400]!, 22, false),
                                     contentPadding: const EdgeInsets.all(9),
                                     border: InputBorder.none,
                                   ),
@@ -264,7 +264,7 @@ class _DesktopTradeState extends State<DesktopTrade> {
                               child: Text(
                                 "Slippage tolerance:",
                                 style: textStyle(
-                                    Colors.grey[400]!, 12, false, false),
+                                    Colors.grey[400]!, 12, false),
                               ),
                             ),
                             Container(
@@ -272,7 +272,7 @@ class _DesktopTradeState extends State<DesktopTrade> {
                               child: Text(
                                 "2%",
                                 style: textStyle(
-                                    Colors.grey[400]!, 12, false, false),
+                                    Colors.grey[400]!, 12, false),
                               ),
                             ),
                           ],
@@ -330,7 +330,7 @@ class _DesktopTradeState extends State<DesktopTrade> {
                             alignment: Alignment.centerLeft,
                             child: Text(
                               token.ticker,
-                              style: textStyle(Colors.white, 14, true, false),
+                              style: textStyle(Colors.white, 14, true),
                             )),
                         Container(
                             width: 125,
@@ -338,7 +338,7 @@ class _DesktopTradeState extends State<DesktopTrade> {
                             child: Text(
                               token.name,
                               style:
-                                  textStyle(Colors.grey[100]!, 9, false, false),
+                                  textStyle(Colors.grey[100]!, 9, false),
                             )),
                       ],
                     ))
@@ -393,7 +393,7 @@ class _DesktopTradeState extends State<DesktopTrade> {
                 ),
                 Container(width: 10),
                 Expanded(
-                  child: Text(tkr, style: textStyle(Colors.white, 16, true, false)),
+                  child: Text(tkr, style: textStyle(Colors.white, 16, true)),
                 ), 
                 Icon(Icons.keyboard_arrow_down, color: Colors.white, size: 25)
               ],
@@ -415,27 +415,14 @@ class _DesktopTradeState extends State<DesktopTrade> {
     showDialog(context: context, builder: (BuildContext context) => fancyDialog);
   }
 
-  TextStyle textStyle(Color color, double size, bool isBold, bool isUline) {
-    if (isBold) if (isUline)
-      return TextStyle(
-          color: color,
-          fontFamily: 'OpenSans',
-          fontSize: size,
-          fontWeight: FontWeight.w400,
-          decoration: TextDecoration.underline);
-    else
+  TextStyle textStyle(Color color, double size, bool isBold) {
+    if (isBold)
       return TextStyle(
         color: color,
         fontFamily: 'OpenSans',
         fontSize: size,
         fontWeight: FontWeight.w400,
       );
-    else if (isUline)
-      return TextStyle(
-          color: color,
-          fontFamily: 'OpenSans',
-          fontSize: size,
-          decoration: TextDecoration.underline);
     else
       return TextStyle(
         color: color,
@@ -450,141 +437,4 @@ class _DesktopTradeState extends State<DesktopTrade> {
         borderRadius: BorderRadius.circular(rad),
         border: Border.all(color: borCol, width: borWid));
   }
-}
-
-class AthleteTokenList extends StatefulWidget {
-  final BuildContext context;
-  final int tknNum;
-  final Widget Function(Token, int) createTokenElement;
-  AthleteTokenList(this.context, this.tknNum, this.createTokenElement);
-
-  @override
-  _AthleteTokenListState createState() => _AthleteTokenListState();
-}
-
-class _AthleteTokenListState extends State<AthleteTokenList> {
-  SwapController swapController = Get.find();
-  double fromAmount = 0.0;
-  double toAmount = 0.0;
-  bool allFarms = true;
-  Token? tkn1;
-  Token? tkn2;
-  List<Token> tokenListFilter = [];
-  int tokenNumber = 0;
-
-  List<Token> tokens = [
-    AXT("AthleteX", "AX", AssetImage('../assets/images/x.jpg')),
-    SXT("SportX", "SX", AssetImage('../assets/images/sx.png')),
-    MATIC("Matic/Polygon", "Matic", AssetImage('../assets/images/matic.png')),
-  ];
-
-  @override
-  void initState() {
-    super.initState();
-    tkn1 = tokens[0];
-    tkn2 = tokens[1];
-
-    tokenNumber = widget.tknNum;
-
-    for (Athlete ath in AthleteList.list)
-      tokens.add(Token(ath.name + " APT", ath.name + " APT",
-          AssetImage('../assets/images/apt.png')));
-
-    tokenListFilter = tokens;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    double _height = MediaQuery.of(context).size.height;
-    return Dialog(
-        backgroundColor: Colors.grey[900],
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(12.0),
-        ),
-        child: Container(
-            width: 400,
-            height: _height * .65,
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.end,
-              children: <Widget>[
-                // column of elements
-                Container(
-                    height: _height * .625,
-                    width: 350,
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                      children: <Widget>[
-                        Row(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            children: <Widget>[
-                              Container(
-                                  height: 30,
-                                  alignment: Alignment.centerLeft,
-                                  child: Text("Token Name",
-                                      style: textStyle(Colors.grey[400]!, 16,
-                                          false, false))),
-                              Container(
-                                  child: TextButton(
-                                onPressed: () => Navigator.pop(context),
-                                child: Icon(Icons.close,
-                                    color: Colors.grey[400], size: 30),
-                              ))
-                            ]),
-                        Container(
-                          child: Divider(thickness: 1, color: Colors.grey[400]),
-                        ),
-                        createSearchBar(),
-                        Container(
-                            height: _height * .625 - 100,
-                            child: ListView.builder(
-                                physics: BouncingScrollPhysics(),
-                                itemCount: tokenListFilter.length,
-                                itemBuilder: (context, index) {
-                                  return widget.createTokenElement(
-                                      tokenListFilter[index], tokenNumber);
-                                }))
-                      ],
-                    ))
-              ],
-            )
-        )
-    );
-  }
-  
-  Widget createSearchBar() {
-    return Container(
-      width: 300,
-      height: 40,
-      decoration: boxDecoration(Colors.grey[900]!, 100, 1, Colors.grey[300]!),
-      child: Row(
-        // crossAxisAlignment: CrossAxisAlignment.start,
-        mainAxisAlignment: MainAxisAlignment.start,
-        children: <Widget>[
-          Container(width: 8),
-          Container(
-            child: Icon(Icons.search, color: Colors.white),
-          ),
-          Container(width: 10),
-          Expanded(
-            child: Container(
-              child: TextFormField(
-                onChanged: (value) {
-                  setState(() {
-                    tokenListFilter = tokens.where((token) => token.ticker.toUpperCase().contains(value.toUpperCase())).toList();
-                  });
-                },
-                decoration: InputDecoration(
-                  border: InputBorder.none,
-                  contentPadding: EdgeInsets.only(bottom: 8.5),
-                  hintText: "Search a name or paste an address",
-                  hintStyle: TextStyle(color: Colors.white),
-                ),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
 }
